backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
crewai==0.201.1
crewai-tools==0.75.0
anthropic==0.39.0
langchain-anthropic==0.2.3
yfinance==0.2.49
python-dotenv==1.0.1
pydantic==2.10.3
pandas==2.2.3